+++ CHANGELOG.md
@@ -5,9 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [0.1.1]
+## [unreleased]
+### Added
+- Build reference, use prebuilt reference
 ### Fixed
 - EDR for 1/1 variants (have ref, but distance should be counted with only alt)
+- Set SD to 0 for GT .
 
 ## [0.1.0]
 ### Added